feat: create a simple Sign API example on top of the WS client
This client implements pairing and session flows:
https://specs.walletconnect.com/2.0/specs/clients/core/pairing
https://specs.walletconnect.com/2.0/specs/clients/sign/session-proposal

Supported actions:
- pairing
- session establishment
- session delete
- ping

This example could be expanded to handle multiple sessions.

Caution:
The purpose of this example is demonstration of core Sign API
functionality. However, it shouldn't be used in production, as might
exhibit some race conditions such as with session deletion, etc...